finalize build.
