+++ lib/index.js
@@ -72,7 +72,6 @@ var WebpackShellPlugin = function () {
     babelHelpers.classCallCheck(this, WebpackShellPlugin);
 
     this.options = validateInput(mergeOptions(options, defaultOptions));
-    console.log(this.options);
   }
 
   babelHelpers.createClass(WebpackShellPlugin, [{
